Move AkkaGrpcClient out of 'internal'
