+++ play-testkit/src/main/scala/akka/grpc/play/api/specs2/ServerGrpcClient.scala
@@ -5,8 +5,9 @@
 package akka.grpc.play.api.specs2
 
 import scala.reflect.ClassTag
-import akka.grpc.internal.{ AkkaGrpcClient, AkkaGrpcClientFactory }
+import akka.grpc.internal.AkkaGrpcClientFactory
 import akka.grpc.play.AkkaGrpcClientHelpers
+import akka.grpc.scaladsl.AkkaGrpcClient
 import play.api.test.{ NewForServer, RunningServer }
 
 /**
