+++ runtime/src/main/scala/akka/grpc/scaladsl/AkkaGrpcClient.scala
@@ -1,0 +1,17 @@
+/*
+ * Copyright (C) 2018 Lightbend Inc. <https://www.lightbend.com>
+ */
+
+package akka.grpc.scaladsl
+
+import scala.concurrent.Future
+
+import akka.Done
+import akka.annotation.DoNotInherit
+
+/** Common trait of all generated Akka gRPC clients. Not for user extension. */
+@DoNotInherit
+trait AkkaGrpcClient {
+  def close(): Future[Done]
+  def closed(): Future[Done]
+}
